separate int test
